feat: StringParser 예외 처리 로직 추가
- rename
