--- backend/src/main/java/com/h2o/h2oServer/global/util/StringParser.java
@@ -1,19 +1,0 @@
-package com.h2o.h2oServer.global.util;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class StringParser {
-    public static List<Long> parseToLongList(String string) {
-        return Arrays.stream(string.split(","))
-                .map(Long::parseLong)
-                .collect(Collectors.toList());
-    }
-
-    public static String parseToString(List<Long> list) {
-        return list.stream()
-                .map(String::valueOf)
-                .collect(Collectors.joining(","));
-    }
-}
